refactor: Extract transaction management from Database into TransactionManager
Addresses issue #1028 by separating transaction concerns from the Database struct.

Changes:
- Created new `transactions.rs` module with:
  - TransactionChange, Savepoint, TransactionState types
  - TransactionManager struct handling all transaction lifecycle
- Updated Database to delegate transaction operations to TransactionManager
- Reduced database.rs from 528 to 394 lines (134 lines removed)
- Maintained backward-compatible public API
- All existing tests pass including transaction-specific tests

Benefits:
- Better separation of concerns (Database focuses on storage, TransactionManager on transactions)
- Clearer ownership of transaction state
- Easier to test transaction logic independently
- More maintainable codebase structure

The Database struct still provides the same transaction methods but now delegates
to TransactionManager, preserving the existing API for consumers.
